requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.1
django-baton==2.8.1
django-bootstrap5==23.4
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
exponent-server-sdk==2.0.0
fonttools==4.47.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
ua-parser==0.18.0
urllib3==2.1.0
user-agents==2.2.0